+++ package.json
@@ -8,7 +8,7 @@
     "build": "turbo run build",
     "clean": "turbo run clean",
     "dev": "turbo run dev --no-cache --parallel --continue",
-    "dev:listen": "turbo run dev --filter=listen --filter=ui --no-cache --parallel --continue",
+    "dev:listen": "turbo run dev --filter=core --filter=listen --filter=ui --no-cache --parallel --continue",
     "dev:admin": "turbo run dev --filter=admin --filter=core --no-cache --parallel --continue",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "lint": "turbo run lint",
