+++ packages/core/src/index.ts
@@ -1,1 +1,4 @@
+import { hooks as SHooks } from "./universal";
+
 export * as ListenCore from "./listen";
+export { SHooks };
