+++ packages/core/src/universal/hooks/index.ts
@@ -1,0 +1,3 @@
+import useSAudioPlayer from "./useSAudioPlayer";
+
+export { useSAudioPlayer };
