+++ packages/ui/package.json
@@ -38,6 +38,7 @@
     "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.10.9",
     "@mui/material": "^5.10.12",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "core": "workspace:*"
   }
 }
